Fix circular import with adding new files
