+++ static/app/views/dashboards/widgetCard/widgetCardContextMenuContainer.tsx
@@ -1,0 +1,6 @@
+import styled from '@emotion/styled';
+
+export const WidgetCardContextMenuContainer = styled('div')`
+  opacity: 1;
+  transition: opacity 0.1s;
+`;
